+++ Backend/lession/src/lession.Application/DTOs/SanPham/UpdateSanPhamDto.cs
@@ -5,9 +5,9 @@
         public string MaSanPham { get; set; } = string.Empty;
         public string TenSanPham { get; set; } = string.Empty;
         public string MoTa { get; set; } = string.Empty;
-        public decimal GiaBan { get; set; }
+        public float GiaBan { get; set; }
         public bool? Active { get; set; }
-        public int SoLuongTon { get; set; }
+        public float SoLuongTon { get; set; }
         public string DonViTinh { get; set; } = string.Empty;
         // Ngày cập nhật sẽ
     }
